# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SM Torneio API",
    "description": "API do Torneio Recreativo Secos & Molhados — Tijuca Tênis Clube.\n\nTodas as rotas estão sob o prefixo `/api`. As rotas *tournament-scoped* usam o slug do torneio (`2026`, `demo`, etc) no caminho.",
    "version": "1.1.0"
  },
  "servers": [
    { "url": "http://localhost:3001/api", "description": "Desenvolvimento local" },
    { "url": "https://seu-dominio.up.railway.app/api", "description": "Produção (Railway)" }
  ],
  "paths": {
...
    "/tournaments/{slug}/standings/{group}": {
      "get": {
        "summary": "Classificação atual de um grupo",
        "tags": ["Standings"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          {
            "name": "group",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["F", "M"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista ordenada de standings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Standing" }
                }
              }
            }
          }
        }
      }
    },
    "/tournaments/{slug}/standings/simulate": {
      "post": {
        "summary": "Simular o efeito de uma etapa no ranking",
        "tags": ["Standings"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SimulationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranking simulado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SimulatedStanding" }
                }
              }
            }
          }
        }
      }
    },
    "/tournaments/{slug}/standings/{participantId}": {
      "put": {
        "summary": "Atualizar manualmente os números de classificação de uma participante",
        "description": "Permite ao admin forçar números de pontos, participações e estatísticas em uma linha de classificação. Útil para correções manuais.",
        "tags": ["Standings"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "participantId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StandingUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standing atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Standing" }
              }
            }
          },
          "404": { "description": "Standing não encontrada" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Standing": {
        "type": "object",
        "properties": {
          "position": { "type": "integer" },
          "participantId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "group": { "type": "string", "enum": ["F", "M"] },
          "pointsValid": { "type": "integer" },
          "pointsRaw": { "type": "integer" },
          "pointsDiscard": { "type": "integer" },
          "pointsBonus": { "type": "integer" },
          "pointsPenalty": { "type": "integer" },
          "roundsPlayed": { "type": "integer" }
        }
      },
      "SimulationRequest": {
        "type": "object",
        "properties": {
          "group": { "type": "string", "enum": ["F", "M"] },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "participantId": { "type": "string" },
                "present": { "type": "boolean" },
                "absentReason": {
                  "type": "string",
                  "enum": ["NONE", "FALTA", "SORTEIO", "SORTEIO_VOLUNTARIA"]
                },
                "position": { "type": "integer", "nullable": true }
              },
              "required": ["participantId"]
            }
          }
        },
        "required": ["group", "results"]
      },
      "SimulatedStanding": {
        "allOf": [
          { "$ref": "#/components/schemas/Standing" },
          {
            "type": "object",
            "properties": {
              "oldPosition": { "type": "integer", "nullable": true },
              "positionDelta": { "type": "integer" },
              "oldPointsValid": { "type": "integer", "nullable": true }
            }
          }
        ]
      },
...
      "StandingUpdateRequest": {
        "type": "object",
        "description": "Qualquer campo fornecido será atualizado na linha de classificação correspondente.",
        "properties": {
          "pointsRaw": { "type": "integer" },
          "pointsDiscard": { "type": "integer" },
          "pointsBonus": { "type": "integer" },
          "pointsPenalty": { "type": "integer" },
          "pointsValid": { "type": "integer" },
          "roundsPlayed": { "type": "integer" },
          "firstPlaces": { "type": "integer" },
          "secondPlaces": { "type": "integer" },
          "thirdPlaces": { "type": "integer" },
          "fourthPlaces": { "type": "integer" },
          "fifthPlaces": { "type": "integer" },
          "sixthPlaces": { "type": "integer" },
          "seventhPlaces": { "type": "integer" },
          "setsWon": { "type": "integer" },
          "setsLost": { "type": "integer" },
          "gamesWon": { "type": "integer" },
          "gamesLost": { "type": "integer" }
        }
      }
    }
  }
}